Documented `Entry` and added additional features. Used `pyglet.app.run`
On discord a couple of weeks ago, we found out that the fps was capped to 60. Using pyglet's run function, we were able to speed things up. The `Entry` was heavily documented, and many features, like copy and paste, select all, and cut were added. There are still some problems with the `Entry`.